# docs/contracts/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Cerberus API",
    "description": "Cerberus encrypted vault API. Identity authentication and vault access are separate.",
    "version": "v1"
  },
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationOutputDataOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationOutputDataOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationOutputDataOutput"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/2fa/verify": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyChallengeCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyChallengeCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyChallengeCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationOutputDataOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationOutputDataOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationOutputDataOutput"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessOutput"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthenticationAccount": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "revision": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
      "AuthenticationOutput": {
        "required": [
          "identity"
        ],
        "type": "object",
        "properties": {
          "identity": {
            "$ref": "#/components/schemas/HeimdallLogin"
          },
          "account": {
            "$ref": "#/components/schemas/AuthenticationAccount"
          }
        },
        "additionalProperties": false
      },
      "AuthenticationOutputDataOutput": {
        "type": "object",
        "properties": {
          "messages": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "success": {
            "type": "boolean",
            "readOnly": true
          },
          "data": {
            "$ref": "#/components/schemas/AuthenticationOutput"
          }
        },
        "additionalProperties": false
      },
...
      "HeimdallLogin": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "nullable": true
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "emailVerified": {
            "type": "boolean",
            "nullable": true
          },
          "requiresTwoFactor": {
            "type": "boolean"
          },
          "challengeToken": {
            "type": "string",
            "nullable": true
          },
          "availableMethods": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "LoginCommand": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProcessOutput": {
        "type": "object",
        "properties": {
          "messages": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "success": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "VerifyChallengeCommand": {
        "type": "object",
        "properties": {
          "challengeToken": {
            "type": "string",
            "nullable": true
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "recoveryCode": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  },
...
}